### 2017-11-20 #466 @whedon
[PRE REVIEW]: An easy-to-use p5.js 3D object picker for visual artists
**Submitting author:** @physicsdavid (<a href="http://orcid.org/0000-0002-8321-2062">David Harris</a>)
**Repository:** <a href="https://github.com/physicsdavid/mPicker" target ="_blank">https://github.com/physicsdavid/mPicker</a>
**Version:** v1.0
