### 2017-11-20 #466 @physicsdavid
Of the existing reviewer list, perhaps @fil might be most appropriate but it's unclear given that this submission is outside the scope of most JOSS submissions. In the pre-submission query, @kyleniemeyer was open to a submission of this kind. I believe you also had some discussions with the editorial team so you might have a clearer idea.
