### 2017-11-21 #466 @arfon
Thanks @physicsdavid.

@Fil - would you be willing to review this submission for JOSS?
