### 2017-11-30 #466 @arfon
@Fil - please head over to #475 to carry out the actual review.
