### 2017-11-30 #475 @Fil
Hello David (@physicsdavid)

After a first look at the requirements, a few of them are found insufficient:
- version: needs a corresponding *release*
- statement of need (both in the software and the paper): needs a better explanation of why it's difficult to do this object picking in native p5 (maybe a link to an example of how people do it manually)
- Installation instructions: please write out the dependencies. From the example file we see that you include p5 and two addons, are the addons necessary?
- example usage: the example doesn't work out of the box (I had to fix the path to ../mPicker/mPicker.js); it is far from a real-world example (but that's optional); the kitten is adorable
- functionality documentation: the documentation doesn't seem to hold on its own and needs to be read alongside p5 documentation, please at minimum provide a link to relevant page(s) on p5 regarding the intention and parameters of each function
- automated tests: none
- community guidelines: absent

I might be wrong on some of these issues, please don't hesitate to say it. And I will review again when they are fixed.

Though I'm not a p5 user myself (yet), it looks like the use case is indeed invaluable, and worth working to have this paper published in JOSS.
