### 2017-12-01 2a76ffb
Improved documentation
