### 2017-12-01 2fac1f0
Revised paper
More detailed statement of need. Added subheads.
