### 2017-12-01 bf9a286
Improved documentation
Added dependencies. Added link to relevant p5.js reference guide.
