+++ CONTRIBUTING.md
@@ -1,0 +1,11 @@
+# Contributing to mPicker
+
+We welcome contributions of all kinds including improvements to the core code,
+example, documentation, or bug reports, via the usual GitHub processes.
+
+## Repository structure
+
+This repository is structured with the core code as a single javascript file
+in `mPicker/`. Example code is a zipped package in `example/` in the standard p5.js
+format including index.html and sketch.js with mPicker.js and additional asset
+files.
