### 2017-12-01 #475 @physicsdavid
Thanks, Philippe (@Fil). These are very helpful comments. I believe I have addressed your issues with the following notes:
- dependencies/example: Although the two p5 addons are not strictly necessary, the standard p5.js template package includes them as part of the index.html file as they are commonly used in many p5.js applications and their inclusion in index.html appears in most p5.js applications. I have also packaged the example as a .zip as this is the way that many p5.js applications are presented, including those downloaded from the p5.js web editor at http://alpha.editor.p5js.org/
- automated tests: given the nature of the p5.js environment, there does not seem to be an appropriate automated test for this code, but the example code includes an implementation of the core functionality of this code so could be used as a test for future changes.

Feel free to ask for further revisions. Thanks again.
