### 2017-12-01 #475 @Fil
Thank you @physicsdavid for all the changes. They do indeed address all the issues!

(The only caveat is I'm not sure I understand why you have made a zip file of the example — it looked more directly accessible and readable as plain files one could look up on github or open in a browser. But it's not an issue for this review.)

@arfon my review is complete and positive for publication.
