### 2017-12-01 #475 @physicsdavid
Thanks, @Fil!

One of the common ways to distribute p5.js applications is via a zip file that contains the set of index.html, sketch.js, plus other assets as required. If you develop a sketch in the online editor (as many of my students do, for example), then if you want to download the sketch, you get a pre-packaged zip with everything in the right folder structure, etc. It's certainly not necessary and I could change it back. One other advantage is that many artists beginning to work with code won't be particularly familiar with github and will likely find it easier to just download one zip file with the example than deal with the files separately. But I'm open to whichever option seems better!
