### 2017-12-01 074abb3
Re-organized example code structure
