### 2017-12-01 #475 @physicsdavid
@Fil, I've reconsidered and changed the file structure of the example to make it consistent with the release as the downloadable version with the example files easily readable as separate files within GitHub. I'm still getting used to appropriate use of GitHub!
