### 2017-12-04 #475 @arfon
@Fil @physicsdavid - where are we at with this review? Is there still more to be done here?
