### 2017-12-04 #475 @Fil
Yes @arfon my review is complete and positive. Thank you @physicsdavid for
the most recent structure changes, makes it even easier to read the example.
